Cloisonne les donnees par entreprise
`id_entreprise` existait depuis le premier commit sur presque toutes les tables sans que rien ne le
renseigne ni ne filtre dessus. Deux entreprises partageant cette base voyaient les articles, les
clients, les ventes et les factures l'une de l'autre — et pouvaient les modifier.

Le principe
- L'entreprise vient du compte connecte, jamais de la requete. Un idEntreprise envoye dans le corps
  est ignore : ce serait une invitation a ecrire chez le voisin. Un test le verifie en pretendant
  travailler pour l'autre entreprise.
- Les listes ne rendent que l'entreprise de l'appelant. Les recherches par code sont cloisonnees
  des la requete : deux entreprises peuvent employer le meme code d'article, et rien ne l'interdit.
- Lire, modifier ou supprimer la donnee d'une autre entreprise rend un 404 et non un 403. Repondre
  « interdit » confirmerait son existence, et permettrait de deviner ce que le voisin possede en
  essayant des identifiants.
- ROLE_SUPER_ADMIN regarde au-dela d'une entreprise : c'est l'editeur, celui qui les cree, pas le
  gerant. ROLE_ADMIN administre la sienne.

Tout cela se decide dans une seule classe, Cloisonnement, pour que la reponse soit la meme partout.
Elle distingue trois « pas d'entreprise » qui ne veulent pas dire la meme chose : le
super-administrateur, un compte non rattache — qui ne voit que les donnees sans entreprise — et un
appel hors authentification, traitement interne ou test, ou rien n'est filtre. Toutes les routes
HTTP exigeant un compte, ce dernier cas ne se presente pas a travers l'API.

Applique a : articles, categories, clients, fournisseurs, ventes, factures, mouvements de stock et
les deux sortes de commandes. Les mouvements heritent de l'entreprise de leur article, verifie
juste avant — la requete ne decide de rien.

Au passage, mvt_stk.id_entreprise passe de integer a bigint : seule table a s'ecarter du type
commun, et un cloisonnement qui compare des identifiants ne peut pas vivre avec deux types.

Tests : 91 (+8), deux entreprises ouvertes cote a cote.

+++ src/main/java/com/jumpy/tech/gestionstock/gestiondestock/dto/MvtStkDto.java
@@ -19,7 +19,7 @@ public class MvtStkDto {
     private ArticleDto article;
     private TypeMvtStk typeMvt;
     private MotifMvtStk motif;
-    private Integer idEntreprise;
+    private Long idEntreprise;
 
     public static MvtStkDto fromEntity(MvtStk mvtStk) {
         if (mvtStk == null) {
